Fix MSVC compiler warnings treated as errors
- Logger.H: Remove unused exception variable in catch block
- KnownBytes.C: Use static_assert for compile-time size checks (fixes C4127)
- SensorsMap.C: Add explicit cast for size_t to int conversion (fixes C4267)
- Sensors.C: Rename shadowed variable 'is' to 'tokenStart' (fixes C4456)
- CMakeLists.txt: Add _CRT_SECURE_NO_WARNINGS for strerror deprecation (fixes C4996)

+++ src/SensorsMap.C
@@ -97,8 +97,8 @@ SensorsMap::setUpForData()
         if (nt != names.end()) { // Already known
           sensor.index(nt->second);
         } else { // Not seen yet
-          sensor.index(names.size());
-          names.insert(std::make_pair(sensor.name(), names.size()));
+          sensor.index(static_cast<int>(names.size()));
+          names.insert(std::make_pair(sensor.name(), static_cast<int>(names.size())));
           mAllSensors.insert(sensor);
         }
       }
